// src/app/features/employees/models/employee.model.ts
export interface Employee {
  id: number;
  employeeId: number;
  nameAr: string;
  nameEn: string;
  email: string;
  password: string;
  job: string;
  companyId: number;
  status: boolean;
  department_id_random: number;
  department_name_ar: string;
  department_name_en: string;
  address: string | null;
  nationality: string;
  national_id: string | null;
  marital_status: number | null;
  marital_status_name_ar: string | null;
  marital_status_name_en: string | null;
  company_emp_code: string | null;
  system_emp_code: string | null;
  gender: number;
  gender_name_ar: string | null;
  gender_name_en: string | null;
  birthdate: string;
  str_birthdate: string | null;
  joining_date: string;
  str_joining_date: string | null;
  direct_manager_id: number | null;
  direct_manager_name_ar: string | null;
  direct_manager_name_en: string | null;
  employment_status: number;
  employment_status_name_ar: string;
  employment_status_name_en: string;
  emp_image: string | null;
  phone_no: string | null;
  roleNameAr: string;
  roleNameEn: string;
  employment_id: number | null;
  employment_code: string | null;
  roleId: number;
  emp_session_id: string | null;
  decrypted_session_id: string | null;
  job_type: number | null;
  job_type_name_ar: string | null;
  job_type_name_en: string | null;
  shift_type: number | null;
  shift_type_name_ar: string | null;
  shift_type_name_en: string | null;
}